requirements.txt
blake3
argon2-cffi
pycryptodome